test: split billing renewal execution tests
Changed:
Split the billing renewal execution tests into focused files and moved shared setup into local test helper modules.

Benefit:
Billing renewal coverage is easier to review and maintain without changing production behavior.
